requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
et-xmlfile==1.1.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
networkx==2.8.8
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
wrapt==1.14.1
